♻️ hook に切り分け
